Update: 'SDK' finance resources folders;
