+++ packages/sdk/src/microservices/finance/index.js
@@ -1,0 +1,26 @@
+/**
+ * Base
+ */
+import { defaultSettings, request } from './base.js';
+
+/**
+ * Sub Modules
+ */
+import producers from './producers.js';
+import transfers from './transfers.js';
+
+/**
+ * Reference
+ */
+const finance = {
+    defaultSettings,
+    request,
+
+    producers,
+    transfers,
+};
+
+/**
+ * Exporting
+ */
+export default finance;
