+++ packages/sdk/src/microservices/shop/index.js
@@ -1,0 +1,17 @@
+/**
+ * Base
+ */
+import { defaultSettings, request } from './base.js';
+
+/**
+ * Reference
+ */
+const shop = {
+    defaultSettings,
+    request,
+};
+
+/**
+ * Exporting
+ */
+export default shop;
