+++ packages/sdk/src/microservices/tickets/index.js
@@ -1,39 +1,8 @@
 /**
  * Base
  */
-import {
-    get,
-    generic,
-} from '../request/request.js';
-
-/**
- * Get Microservice Default Settings
- *
- * @param {object} settings
- *
- * @returns {object}
- */
-function defaultSettings(settings = {}) {
-    return {
-        microservice: 'ticket',
-        ...settings,
-    };
-}
-
-/**
- * Generic MS Requests
- *
- * @param {string} path
- * @param {object} settings
- *
- * @returns {Promise}
- */
-function request(
-    path,
-    settings
-) {
-    return generic(path, defaultSettings(settings));
-}
+import { get } from '../../request/request.js';
+import { defaultSettings, request } from './base.js';
 
 /**
  * Tickets Microservice Getter
